Refactor: separate move funds action saga
